+++ app/components/files_search/file_options/header.tsx
