+++ app/components/files_search/file_options/icon.tsx
