+++ app/components/files_search/file_options/mobile_options.tsx
