+++ app/components/files_search/file_options/tablet_options.tsx
